refactor(fe): pR 수정사항 및 모달 컴포넌트 불필요한 코드 제거
