+++ fe/src/components/modules/Modal/AccountInfoModal.tsx
@@ -4,14 +4,13 @@ import { ReactComponent as LogoutSVG } from 'assets/svg/logout.svg';
 import { Link } from 'react-router-dom';
 
 interface UserInfo  {
-  accountModalRef?: React.RefObject<HTMLDivElement>;
   userName: string; // 이름은 Header의 Prop으로 받아오는 계정 이름을 이용할 예정
 }
 
 //  onClose의 타입은 테스트가 끝나면 알맞는 바꾸도록하자
-const AccountInfoModal = ({ userName, accountModalRef }: UserInfo) => {
+const AccountInfoModal = ({ userName}: UserInfo) => {
   return (
-    <div ref={accountModalRef} className="modal_info_box">
+    <div className="modal_info_box">
       <p className="username text">{userName}</p>
       <hr />
         <Link to="mypage" className="info_link_area">
