+++ fe/src/pages/Login.tsx
@@ -1,19 +1,10 @@
 import Nav_Login from 'components/login/Naver_Login'
-import getParticles from 'hooks/loginPageSnowAnime';
-import { useEffect } from 'react';
 
 const Login = () => {
-  useEffect(() => {
-    getParticles();
-  }, []);
   return (
     <>
-      <div className="bg"></div>
-      <div className="login_page flex_center">
-        <Nav_Login />
-      </div>
-      <div className="particles" id="particles" />
-    </>  
+      <Nav_Login />
+    </>
   )
 }
