+++ fe/src/styles/components/login/_login_box.scss
@@ -1,3 +1,49 @@
+.login_page{
+  width: 100%;
+  height: 100vh;
+  z-index: 100;
+}
+.login_page,
+.particles{
+  width: 100%;
+  height: 100vh;
+  position: fixed;
+  top: 0;
+  left: 0;
+}
+.particles{
+  position: absolute;
+  background-color: $logoColor;
+  z-index: 50;
+  transform-style: preserve-3d;
+  overflow: hidden;
+}
+
+@keyframes move {
+  0%{
+    transform: translateY(0);
+    opacity: 0;
+  }
+  10% ,90%{
+    opacity: 1;
+  }
+  100%{
+    transform: translateY(45vw);
+    opacity: 0;
+  }
+}
+@keyframes rotate-container {
+  0%{
+    transform:
+      perspective(1000px)
+    translateZ(0px);
+  }
+  100%{
+    transform:
+      perspective(1000px)
+    translateZ(100px);
+  }
+}
 .login_box{
   background-color: $whiteColor;
   width: 20rem;
